Bit of cleanup and slightly better error printing
